requirements.txt
pandas
pydantic
pyyaml
regex
streamlit
matplotlib
altair